requirements.txt
PyYAML>=6.0.1
pydantic>=2.9.0
python-dotenv>=1.0.1
openai>=1.12.0
anthropic>=0.18.1
colorama>=0.4.6